Some refactoring
Move functions to separated files, trying to make navigation in files/folder easier
