+++ src/helm/addRepo.go
@@ -1,0 +1,5 @@
+package helm
+
+func RepoAdd() {
+	// pass
+}
